requirements/requirements.txt
Django==2.2
celery[redis]==4.3.0
django-environ==0.4.5
django_celery_results==1.0.4
djangorestframework==3.9.2
djangorestframework-simplejwt==4.3.0
django-redis==4.10.0
django-rest-swagger==2.2.0
django-extensions==2.1.6
django-filter==2.1.0
django-cors-headers==2.5.2
ipython==7.4.0
parse==1.12.0
pydash==4.7.4
psycopg2==2.8.2
psycopg2-binary==2.8.2
Pillow==6.0.0
requests==2.21.0
sqlparse==0.3.0
sentry-sdk==0.7.10
uWSGI==2.0.18